feat: Add Largest Odd Number in String

+++ Math/Largest Odd Number in String/java/Solution.java
@@ -1,9 +1,9 @@
 class Solution {
     public String largestOddNumber(String num) {
       
-      int n=Integer.parseInt(num);
+      long n=Long.parseLong(num);
       if(n%2!=0)return num;
-      int max=0;
+      long max=0;
 
 while(n!=0){
     if(n%2!=0){
@@ -12,7 +12,7 @@ while(n!=0){
     }
     n=n/10;
 }
-String s=Integer.toString(max);
+String s=Long.toString(max);
 
       //System.out.println(max);
      return s.equals("0")?"":s;
